feat: change all types from float to double

+++ src/main/java/co/tryterra/terraclient/models/v2/body/GlucoseData.java
@@ -32,7 +32,7 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class GlucoseData {
     @JsonProperty("day_avg_blood_glucose_mg_per_dL")
-    private Float dayAvgBloodGlucoseMgPerDl;
+    private Double dayAvgBloodGlucoseMgPerDl;
     @JsonProperty("blood_glucose_samples")
     private List<GlucoseDataSample> bloodGlucoseSamples;
     @JsonProperty("detailed_blood_glucose_samples")
